+++ CHANGELOG.md
@@ -10,6 +10,13 @@
   * Added characters:
     * Yaoyao
     * Alhaitham
+  * Added materials:
+    * Sand Grease Pupa
+    * Pseudo-Stamens
+  * Added boss:
+    * Setekh Wenut
+  * Added links for new materials for interactive map and embedded map
+  * Farming route for Sand Grease Pupa
 
 #### 1.1.15 December-2022